Rename module 'mutual_debt_simplification' to 'simplification'
